IOS-2282 Remove page service
